ciao-launcher: Move agent_client into ssntp.go
The main.go file is getting a little large and contains lots of different
bits of code that do different things.  This commit moves all the snntp
related code into it's own file, making main.go easier to digest and the
ssntp code easier to test.
